build: Bump ABI version to be newer than any "real" libudev.so.0
If we have a stale copy of udev's libudev.so.0 (from udev 182 or earlier)
in the same directory as this shim, presumably we want the shim to
compare as newer, so that ldconfig(8) will prefer to create a symlink
libudev.so.0 -> libudev.so.0.x.y pointing to this shim.

To achieve that, the shim's ABI version number 0.x.y needs to be
strictly newer than 0.13.1, which was the ABI version of udev 182,
before the SONAME bump to libudev.so.1 in systemd 183.

Having versions compare in a correct order is also helpful for the
Steam Runtime.

Resolves: https://github.com/archlinux/libudev0-shim/issues/4
Signed-off-by: Simon McVittie <smcv@collabora.com>